Add simple session management
I've added session management using github.com/alexedwards/scs

The base use is adding the user email to the session cookie when a
login happens. Once that happens validating the existance of the cookie
is enough to mark the user as logged in and the session active.

Using this the loggin link in the nav bar is now set to logout. A logout
handler was also added. This is not meant to be a comprehensive session
management example but a quick way to test and learn the scs package